% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupUMI4C.R
\name{addGrouping}
\alias{addGrouping}
\title{Add grouping of UMI-4C samples}
\usage{
addGrouping(
  umi4c,
  grouping = "sampleID",
  scales = 5:150,
  normalized = TRUE,
  sd = 2
)
}
\arguments{
\item{grouping}{Name of the column in colData used to merge the samples or
replicates. Default: "condition".}

\item{scales}{Numeric vector containing the scales for calculating the
domainogram.}

\item{normalized}{Logical indicating whether UMI-4C profiles should be
normalized to the \code{ref_umi4c} sample/group. Default: TRUE}

\item{sd}{Stantard deviation for adaptative trend.}
}
\description{
Add grouping of UMI-4C samples
}
