% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialNbinomWaldTestUMI4C.R
\name{addPeaksUMI4C}
\alias{addPeaksUMI4C}
\title{Significant fragment caller}
\usage{
addPeaksUMI4C(dds, zscore_thresh = 2, fdr_thresh = 0.1)
}
\description{
Significant fragment caller
}
