% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenes.R
\name{addStepping}
\alias{addStepping}
\title{Add stepping for plotting genes}
\usage{
addStepping(genesDat, coordinates, mcol.name)
}
\arguments{
\item{genesDat}{GRanges object containing gene information.}

\item{coordinates}{GRanges object with coordinates you want to plot.}
}
\description{
Given a GRanges dataset representing genes, will add an arbitrary value for them to be plotted in
the Y axis without overlapping each other.
}
