% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{align}
\alias{align}
\title{Align fastq filee}
\usage{
align(splited_file, align_dir, threads = 1, bowtie_index, pos_viewpoint,
  filter_bp = 1e+07)
}
\arguments{
\item{align_dir}{Sequence for the restriction enzyme to cut.}

\item{threads}{Number of threads to use in the analysis.}

\item{bowtie_index}{Bowtie index file for the reference genome.}

\item{pos_viewpoint}{Vector consist of chromosome, start and end position of the viewpoint.}

\item{fastq_file}{Fastq file to align.}
}
\description{
Align fastq filee
}
