% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentR.R
\name{alignmentR}
\alias{alignmentR}
\title{umi4Cats alignment}
\usage{
alignmentR(pathVenv, wk_dir, threads, ref_gen, bait_seq, bait_pad, res_e)
}
\arguments{
\item{pathVenv}{UMI4Cats virtual enviroment path}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{threads}{Number of threads to use in the analysis}

\item{ref_gen}{Reference genome in fasta format}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{res_e}{Restriction enzyme sequence}
}
\description{
Align splitted UMI reads agains a pre-defined reference genome using Bowtie2
}
\examples{
\dontrun{
pathVenv = '/imppc/labs/lplab/share/marc/venv/umi4catsVenv'
wk_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/MLH1_ctrl_umi4cats_python'
threads = 1
bait_seq = 'ACCTAGAAGGATATGCGCTTGC'
bait_pad = 'GCGTTAGA'
res_e = "GATC"
cut_pos = 0
ref_gen = '/imppc/labs/lplab/share/marc/refgen/hg19/hg19.fa'

alignmentR(wk_dir, threads, ref_gen, bait_seq, bait_pad, res_e)
}

}
