% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{alignScript}
\alias{alignScript}
\title{Umi4c alignment}
\usage{
alignScript(wk_dir, cores, ref_gen, bowtie2 = "bowtie2")
}
\arguments{
\item{wk_dir}{Working directory where to carry out the UMI-4c analysis.}

\item{cores}{Number of cores to use in the analysis}

\item{ref_gen}{Reference genome in fasta format.}

\item{bowtie2}{Bowtie2 path}
}
\description{
This function aligns the splited reads using a defined reference genome.
}
