% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculteDomainogram.R
\name{calculateDomainogram}
\alias{calculateDomainogram}
\title{Create Domainogram}
\usage{
calculateDomainogram(umi4c, scales = 5:150)
}
\value{
A matrix where the first column represents the fragment end coordinates (start) and the rest represent the
number of UMIs found when using a specific scale.
}
\description{
Using as input the raw UMIs, this package creates a domainogram for the supplied scales.
}
