% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{contactsUMI4C}
\alias{contactsUMI4C}
\title{UMI4C Contacts Processing}
\usage{
contactsUMI4C(
  fastq_dir,
  wk_dir,
  bait_seq,
  bait_pad,
  res_enz,
  cut_pos,
  digested_genome,
  ref_gen,
  threads = 1,
  path_venv,
  fastq_multx = "fastq-multx"
)
}
\arguments{
\item{fastq_dir}{Path of the directory containing the FastQ files (compressed or uncompressed).}

\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{bait_seq}{Character containing the bait primer sequence.}

\item{bait_pad}{Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{cut_pos}{Numeric indicating the nucleotide position where restriction enzyme cuts (zero-based) (for example, for DpnII is 0).}

\item{digested_genome}{Path for the digested genome file generated using the \code{\link{digestGenome}} function.}

\item{ref_gen}{Path for the reference genome to use for the alignment (fasta format).}

\item{threads}{Number of threads to use in the analysis.}

\item{path_venv}{Path for the UMI4Cats virtual environment generated using the \code{\link{installVenv}}.}

\item{fastq_multx}{Path or character for calling the \code{fastq-multx} script to demultiplex reads.}
}
\description{
Using demultiplexed FastQ files as input, performs all necessary steps to end up with a tsv file summarizing the
restriction enzyme fragments and the number of UMIs supporting that specific contact with the viewpoint (bait) of
interest.
}
\details{
This function is a combination of calls to other functions that perform the necessary steps for processing
UMI-4C data.
}
\examples{
\dontrun{
contactsUMI4C(fastq_dir="raw_fastq",
              wk_dir="SOCS1",
              bait_seq="CCCAAATCGCCCAGACCAG",
              bait_pad="GCGCG",
              res_enz="GATC",
              cut_pos=0,
              digested_genome=hg19_dpnii,
              ref_gen="~/data/reference_genomes/hg19/hg19.fa",
              threads=1,
              path_venv="~/venvs/UMI4Cats_venv/",
              fastq_multx="fastq-multx")
}
}
