% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{count}
\alias{count}
\title{Counts umis for a given bam file.}
\usage{
count(
  digested_genome,
  pos_viewpoint,
  filtered_bam_R1,
  filtered_bam_R2,
  count_dir
)
}
\arguments{
\item{digested_genome}{Path for the digested genome file generated using the \code{\link{digestGenome}} function.}

\item{pos_viewpoint}{Vector consist of chromosome, start and end position of the viewpoint.}

\item{filtered_bam_R1}{R1 bam file,}

\item{filtered_bam_R2}{R2 bam file,}

\item{count_dir}{Counter directory.}

\item{bowtie_index}{Bowtie index file for the reference genome.}
}
\description{
Counts umis for a given bam file.
}
