% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{counterUMI4C}
\alias{counterUMI4C}
\title{UMI counting}
\usage{
counterUMI4C(
  wk_dir,
  pos_viewpoint,
  res_enz,
  digested_genome,
  filter_bp = 1e+07
)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{digested_genome}{Path for the digested genome file generated using the \code{\link{digestGenome}} function.}
}
\description{
Algorithm for counting  and collapsing the number of UMIs supporting a specific ligation.
}
\details{
For collapsing different molecules into the same UMI, takes into account the ligation position and
the number of UMI sequence mismatches.
}
\examples{
\dontrun{
counterUMI4C(wk_dir="SOCS1",
            bait_seq="CCCAAATCGCCCAGACCAG",
            bait_pad="GCGCG",
            res_enz="GATC",
            digested_genome=hg19_dpnii,
            ref_gen="~/data/reference_genomes/hg19/hg19.fa")
}
}
