% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{createDomainogram}
\alias{createDomainogram}
\title{Create Domainogram}
\usage{
createDomainogram(UMI4C, scales = 5:150)
}
\arguments{
\item{UMI4C}{UMI4C objected created with the \code{\link{UMI4C}} function.}

\item{scales}{Scales to use for smoothing and plotting the domainogram. Scales indicate how many fragment
ends are merged together to obtain the normalized number of UMIs. Default: 5:150.}
}
\value{
A matrix where the first column represents the fragment end coordinates (start) and the rest represent the
number of UMIs found when using a specific scale.
}
\description{
Using as input the raw UMIs, this package creates a domainogram for the supplied scales.
}
