% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsUMI4C.R
\name{createStatsTable}
\alias{createStatsTable}
\title{Create stats table}
\usage{
createStatsTable(fastq_dir, wk_dir)
}
\arguments{
\item{fastq_dir}{Path of the directory containing the FastQ files (compressed or uncompressed).}

\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}
}
\description{
Create stats table
}
