% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{darken}
\alias{darken}
\title{Darken colors}
\usage{
darken(color, factor = 1.4)
}
\arguments{
\item{color}{Character containing the name or hex value of a color.}

\item{factor}{Numeric representing a factor by which darken the specified color.}
}
\value{
Darkens the provided color by the provided factor.
}
\description{
Darken colors
}
\examples{
darken("blue", factor=1.4)
}
