% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexFastq.R
\name{demultiplexFastq}
\alias{demultiplexFastq}
\title{Demultiplex FASTQ files using fastq-multx}
\usage{
demultiplexFastq(barcodes, fastq, type = "paired", out_path = "raw_fastq")
}
\arguments{
\item{barcodes}{Dataframe with "name of sample" and "barcode" for every sample to demultiplex}

\item{fastq}{Name of the fastq to demultiplex. When paired-end, you only need to provide the name for the "_R1" file. Different pairs should be named "_R1" or "_R2".}

\item{type}{Character indicating wether fastq files are in "single" or "paired" end format. Default is \code{"paired"} format.}

\item{out_path}{Path where to save the demultiplex output. Defaults to a path named \code{raw_fastq} in your working
directory.}
}
\description{
Demultiplex FASTQ files containng different bait information using \link[=https://github.com/brwnj/fastq-multx]{fastq-multx} software
}
\examples{
\dontrun{
barcodes <- data.frame(sample=c("KLK3", "KLK6", "KLK7", "KLK10"),
                       barcode=c("ATGGTCTGGGCGCTGTCTTG",
                                 "TATTCTTCCTCAGCCCACATCTT",
                                 "GGATGAAGATTTTGGAGCCCAGC",
                                 "GGGCGGGGATTGAACGC"))

demultiplexFastq(barcodes=barcodes,
                 fastq="~/samples/sample_1_R1.fastq.gz",
                 type="paired",
}

}
