% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialUMI4C.R
\name{deseq2UMI4c}
\alias{deseq2UMI4c}
\title{Differential UMI4C contacts using Fisher's Exact test}
\usage{
deseq2UMI4c(umi4c, design = ~condition, normalized = TRUE,
  padj_method = "fdr", ...)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{design}{A \code{formula} or \code{matrix}. The formula expresses how the counts for
each fragment end depend on the variables in \code{colData}. See  \code{\link[DESeq2]{DESeqDataSet}}}

\item{normalized}{Logical indicating if the function shoult return normalized or raw UMI counts.}

\item{padj_method}{The method to use for adjusting p-values, see \code{\link[stats]{p.adjust}}}

\item{...}{Other arguments to be passed to \code{\link[DESeq2]{DESeq}} function.}
}
\description{
Differential UMI4C contacts using Fisher's Exact test
}
