% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialNbinomWaldTestUMI4C.R
\name{differentialNbinomWaldTestUMI4C}
\alias{differentialNbinomWaldTestUMI4C}
\title{Differential UMI4C contacts using DESeq2 Wald Test}
\usage{
differentialNbinomWaldTestUMI4C(
  umi4c,
  design = ~condition,
  normalized = TRUE,
  padj_method = "fdr",
  query_regions = NULL,
  padj_threshold = 0.05,
  penalty = 0.1,
  alpha = 20
)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the
\code{UMI4C} constructor.}

\item{design}{A \code{formula} or \code{matrix}. The formula expresses how
the counts for each fragment end depend on the variables in \code{colData}.
See  \code{\link[DESeq2]{DESeqDataSet}}. By default ~condition}

\item{normalized}{Logical indicating if the function shoult return normalized
or raw UMI counts. By default TRUE.}

\item{padj_method}{The method to use for adjusting p-values, see
\code{\link[stats]{p.adjust}}. By default fdr.}

\item{query_regions}{\code{GRanges} object or \code{data.frame} containing
the coordinates of the genomic regions you want to use to perform the
analysis in specific genomic intervals. By default NULL.}

\item{padj_threshold}{Numeric indicating the adjusted p-value threshold to
use to define significant differential contacts. By default 0.05.}

\item{penalty}{Amount of smoothing to be applied to the estimated functional parameter. By default 0.1.}

\item{alpha}{Approximate number of fragments desired for every basis function of the B-spline basis. 
floor((max(number of fragments))/alpha) is passed to create.bspline.basis as nbasis 
argument (if this value is smaller than 4, 4 is the default value). By default 20.}
}
\value{
UMI4c object with the DESeq2 Wald Test results,
}
\description{
Takes an UMI4c object and infers the differences between conditions of the smooth monotone fit values
using DESeq2 Wald Test from \code{DESeq2} package.
}
\details{
This function fits the signal trend of a variance stabilized count values using 
a symmetric monotone fit for the distance dependency. Then scales the raw counts across the samples to
obtin normalizaed factors. Finally, it detects differences between conditions applying the DESeq2 Wald Test.
}
