% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestGenome.R
\name{digestGenome}
\alias{digestGenome}
\title{Digest reference genome}
\usage{
digestGenome(cut_seq_5p, cut_seq_3p, name_RE, refgen,
  out_path = "genome_digest")
}
\arguments{
\item{cut_seq_5p}{5' restriction enzyme cut sequence. For example, for DpnII |GATC \code{cut_seq_5p=""}.}

\item{cut_seq_3p}{3' restriction enzyme cut sequence. For example, for DpnII |GATC \code{cut_seq_3p="GATC"}.}

\item{name_RE}{Restriction enzyme name .}

\item{refgen}{A BSgenome object of the reference genome.}

\item{out_path}{Output path where to save the genomic track. The default is a directory named \code{genome_digest/}
created in your working directory.}
}
\description{
Performs an \emph{in silico} digestion of a given reference genome using a given restriction enzyme sequence.
}
\examples{
\dontrun{
library(BSgenome.Hsapiens.UCSC.hg19)
refgen <- BSgenome.Hsapiens.UCSC.hg19
output <- '~/refgen/genomicTracksR'
cs5p <- ""
cs3p <- "GATC"

hg19_dpnii <- digestGenome(cut_seq_5p=cs5p,
                           cut_seq_3p=cs3p,
                           name_RE="DpnII",
                           refgen=refgen,
                           out_path=output)
}
}
