% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsUMI4C.R
\name{.getSummaryBam}
\alias{.getSummaryBam}
\title{Summarize BAM file}
\usage{
.getSummaryBam(bam_file, mapped = TRUE, secondary = FALSE)
}
\arguments{
\item{bam_file}{Path for the bam file.}

\item{mapped}{Logical indicating whether to extract mapped reads.}

\item{secondary}{Logical indicating whether to extract secondary aligned reads.}
}
\description{
Get summary of interesting bam statistics
}
