% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMI4Cats.R
\docType{data}
\name{ex_ciita_umi4c}
\alias{ex_ciita_umi4c}
\title{Contacts with CIITA promoter}
\format{
A UMI4C object from this package.
}
\source{
See inst/script/CIITA_process_example.R to see the code use for generating
the UMI4C object.
}
\usage{
data("ex_ciita_umi4c")
}
\description{
An example UMI4C object showing the contacts with a viewpoint located at 
the CIITA gene promoter.
}
\keyword{datasets}
