% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqMultxR.R
\name{fastqMultxR}
\alias{fastqMultxR}
\title{Demultiplex using fastq-multx}
\usage{
fastqMultxR(dfBarcodes, fastqFile, type, outPath)
}
\arguments{
\item{dfBarcodes}{Dataframe with "name of sample" and "barcode" for every sample to demultiplex}

\item{fastqFile}{Fastq to demultiplex.
In paired-end file used should be R1 and following the format "file_R1.*"}

\item{type}{"single" or "paired" format}

\item{outPath}{Path where to save the demultiplex output}
}
\description{
Demultiplex samples using fastq-multx software
}
\examples{
barcodeM <- c('KLK3',	'ATGGTCTGGGCGCTGTCTTG',
             'KLK6',	'TATTCTTCCTCAGCCCACATCTT',
             'KLK7',	'GGATGAAGATTTTGGAGCCCAGC',
             'KLK10',	'GGGCGGGGATTGAACGC')

barcodeM <- t(matrix(barcodeM, nrow = 2))
dfBarcodes <- data.frame(barcodeM)

colnames(dfBarcodes) <- c('sample', 'barcode')
fastqFile <- '/home/labs/lplab/msubirana/Documents/IGTP/kallikreins/raw/fastq/umi4c/CTTGTA_R1.fastq.gz'
type <- 'paired'
outPath <- '/home/labs/lplab/msubirana/Documents/IGTP/kallikreins/raw/fastq'

fastqMultxR(dfBarcodes,
           fastqFile,
           type,
           outPath)
}
