% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{filterRaw}
\alias{filterRaw}
\title{Filter raw reads}
\usage{
filterRaw(input, bait_coordinates, bait_exclusion = 3000,
  bait_upstream = 5e+05, bait_downstream = 5e+05)
}
\arguments{
\item{input}{A two-column data.frame where the first column represents the fragend and the second the number of UMIs.}

\item{bait_coordinates}{GRanges object with the coordinates of the bait used for the experiment.}

\item{bait_exclusion}{Region around the bait (in bp) to be excluded from the analysis. Default: 3kb.}

\item{bait_upstream}{Number of bp upstream of the bait to use for the analysis. Default: 500kb.}

\item{bait_downstream}{Number of bp downstream of the bait to use for the analysis. Default: 500kb.}
}
\value{
A data.frame with two columns corresponding to the fragment end (\emph{coord}) and the number of
UMIs supporting the article at that fragment end (\emph{UMIs}).
}
\description{
Filters out all fragment ends around the bait and outside the window of interest.
}
