% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialUMI4C.R
\name{fisherUMI4C}
\alias{fisherUMI4C}
\title{Differential UMI4C contacts using Fisher's Exact test}
\usage{
fisherUMI4C(umi4c, grouping = "condition", query_regions,
  resize = NULL, padj_method = "fdr")
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{grouping}{Name of the \code{colData(UMI4C)} column that will be used for grouping the UMI4C profiles
and performing differential analyses.}

\item{query_regions}{\code{GenomicRanges} object containing the coordinates for the regions where to
perform the differential analysis.}

\item{resize}{Width in base pairs for resizing the query regions. Default: no resizing.}

\item{padj_method}{Method for adjusting p-values. See \code{\link[stats]{p.adjust}} for the different methods.}
}
\description{
Differential UMI4C contacts using Fisher's Exact test
}
