% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{formatPlotsUMI4C}
\alias{formatPlotsUMI4C}
\title{Format plots for UMI4C}
\usage{
formatPlotsUMI4C(plot_list, font_size)
}
\arguments{
\item{plot_list}{List of plots generated by  \code{\link{plotUMI4C}}}

\item{font_size}{Base font size to use for the UMI4C plot. Default: 14.}
}
\value{
Given a named plot_list and considering the number and type of included
plots, formats their axes accordingly to show the final UMI4C plot.
}
\description{
Format plots for UMI4C
}
