% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicTrackGenerator.R
\name{genomicTrackGenerator}
\alias{genomicTrackGenerator}
\title{Genomic track generator}
\usage{
genomicTrackGenerator <- function(cp5p, cs3p, nameRe, refgen, output)
}
\arguments{
\item{cs5p}{5' restriction enzyme cut sequence. In the GA|TC case "GA".}

\item{cs3p}{3' restriction enzyme cut sequence. In the GA|TC case "TC"}

\item{nameRe}{Restriction enzyme name .}

\item{refgen}{Reference genome used for the generation of the genomic tracks.
Choose from BSgenome, for listing the available once use "available.genomes()"}

\item{outPath}{Output path where to save the genomic track.}
}
\description{
In silico digestion of a reference genome
}
