% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processComparisonUMI4C.R
\name{getDomainogramComparison}
\alias{getDomainogramComparison}
\title{Get comparative domainogram}
\usage{
getDomainogramComparison(umi4c_obj, ref_umi4c_obj)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ref_umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile to use as reference.
\code{umi4c_obj} will be scaled to the \code{ref_umi4c_obj}.}
}
\description{
Will return values of a comparative domainogram representing the differences between log2 UMIs.
}
