% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{getFactors}
\alias{getFactors}
\title{Get factors fro plotting}
\usage{
getFactors(umi4c, grouping = NULL)
}
\arguments{
\item{umi4c}{UMI4C object}
}
\value{
Factor vector where the first element is the reference factor.
}
\description{
Get factors fro plotting
}
