% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogsUmi4Cats.R
\name{getLogsUmi4Cats}
\alias{getLogsUmi4Cats}
\title{Get UMI4C stats}
\usage{
getLogsUmi4Cats(fastq_dir, wk_dir, threads)
}
\arguments{
\item{fastq_dir}{Path of the directory containing the FastQ files (compressed or uncompressed).}

\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{threads}{Number of threads to use in the analysis.}
}
\value{
Results are saved in \code{wk_dir/rst.}
}
\description{
Plots stats of UMI4C filtering (based on the presence of bait, pad and restriction enzyme) and the alignment process.
}
\examples{
\dontrun{
fastq_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/MLH1_ctrl_umi4cats_python'
wk_dir = '/imppc/labs/lplab/share/marc/epimutations/raw'
threads = '2'

getLogsUmi4Cats(fastq_dir = fastq_dir, wk_dir = wk_dir, threads = threads)
}

}
