% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNormalizationMatrix.R
\name{getNormalizationMatrix}
\alias{getNormalizationMatrix}
\title{Get normalization matrix}
\usage{
getNormalizationMatrix(umi4c, norm_bins = 10^(3:6),
  post_smooth_win = 50, r_expand = 1.2)
}
\description{
Will return a vector for normalizing domainogram.
}
