% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processComparisonUMI4C.R
\name{getNormalizationVector}
\alias{getNormalizationVector}
\title{Get normalization vector}
\usage{
getNormalizationVector(umi4c_obj, norm_bins = 10^(3:6),
  post_smooth_win = 50, r_expand = 1.2)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{norm_bins}{Bins used to normalize the number of UMIs from \code{umi4c_obj} to \code{umi4c_ref}. Default: 10^(3:6).}
}
\description{
Will return a vector for normalizing domainogram.
}
