% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupUMI4C.R
\name{groupSamplesUMI4C}
\alias{groupSamplesUMI4C}
\title{Group UMI4C samples}
\usage{
groupSamplesUMI4C(umi4c, grouping = "condition")
}
\arguments{
\item{umi4c}{\linkS4class{UMI4C} object as generated by \code{\link{makeUMI4C}}.}

\item{grouping}{Name of the column in colData used to merge the samples or
replicates. Set to NULL for skipping grouping. Default: "condition".}
}
\value{
A grouped \code{UMI4C} object.
}
\description{
Combines UMI4C samples by adding UMIs from \code{assay(umi4c)} to represent
the levels in \code{grouping}.
}
