% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19_gene_annoation_ensemblv75}
\alias{hg19_gene_annoation_ensemblv75}
\title{Gene annotation for hg19 genome}
\format{A GRanges object.}
\usage{
hg19_gene_annoation_ensemblv75
}
\description{
A dataset containing the gene annotation for the hg19 genome build.
}
\keyword{datasets}
