% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{makeUMI4C}
\alias{makeUMI4C}
\title{Make UMI4C object}
\usage{
makeUMI4C(colData, viewpoint_name, normalized = TRUE,
  bait_exclusion = 3000, bait_upstream = 5e+05,
  bait_downstream = 5e+05, scales = 5:150, min_win_factor = 0.02,
  sd = 2)
}
\arguments{
\item{viewpoint_name}{Character indicating the name for the used viewpoint.}

\item{normalized}{Logical indicating whether UMI4C profiles should be normalized to the sample with less UMIs (ref_umi4c). Default: TRUE}

\item{bait_exclusion}{Region around the bait (in bp) to be excluded from the analysis. Default: 3kb.}

\item{bait_upstream}{Number of bp upstream of the bait to use for the analysis. Default: 500kb.}

\item{bait_downstream}{Number of bp downstream of the bait to use for the analysis. Default: 500kb.}

\item{scales}{Numeric vector containing the scales for calculating the domainogram.}

\item{min_win_factor}{Proportion of UMIs that need to be found in a specific window for adaptative trend calcultion}

\item{sd}{Stantard deviation for adaptative trend.}
}
\description{
Make UMI4C object
}
