% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUMI4Cexample.R
\name{makeUMI4Cexample}
\alias{makeUMI4Cexample}
\title{Make UMI4C example object}
\usage{
makeUMI4Cexample(...)
}
\arguments{
\item{...}{Other arguments to be passed to the \code{makeUMI4C} function.}
}
\value{
Creates a UMI4C object to use as an example.
}
\description{
Make UMI4C example object
}
\examples{
umi4c <- makeUMI4Cexample()
}
