% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{mergeUMICounter}
\alias{mergeUMICounter}
\title{Merge UMI4C counts}
\usage{
mergeUMICounter(digested_genome, wk_dir, bait_seq, bait_pad, res_enz, ref_gen)
}
\arguments{
\item{digested_genome}{Path for the digested genome file generated using the \code{\link{digestGenome}} function.}

\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{bait_seq}{Character containing the bait primer sequence.}

\item{bait_pad}{Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{ref_gen}{Path for the reference genome to use for the alignment (fasta format).}
}
\description{
Merge UMI4C counts with digested genome framgment position, returning all fragments and UMIs
10Mb around the viewpoint.
}
\examples{
\dontrun{
mergeUMICounter(wk_dir="SOCS1",
                digested_genome=hg19_dpnii,
                bait_seq="CCCAAATCGCCCAGACCAG",
                bait_pad="GCGCG",
                res_enz="GATC",
                ref_gen="~/data/reference_genomes/hg19/hg19.fa")
}
}
