% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processComparisonUMI4C.R
\name{normalizeDomainogram}
\alias{normalizeDomainogram}
\title{Normalize domainogram to reference}
\usage{
normalizeDomainogram(umi4c_obj, ref_umi4c_obj, norm_bins = 10^(3:6))
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ref_umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile to use as reference.
\code{umi4c_obj} will be scaled to the \code{ref_umi4c_obj}.}

\item{norm_bins}{Bins used to normalize the number of UMIs from \code{umi4c_obj} to \code{umi4c_ref}. Default: 10^(3:6).}
}
\description{
Will normalize the domainogram of a given umi4c_obj to the reference.
}
