% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparisonUMI4C.R
\name{plotComparisonUMI4C}
\alias{plotComparisonUMI4C}
\title{Comparative plot of UMI4C profiles}
\usage{
plotComparisonUMI4C(comp_obj, ymax = NULL, xlim = NULL,
  plot_domainogram = T, protein_coding = T, legend = c(Treatment =
  "darkorange3", Reference = "darkorchid3"), highlight = NULL,
  highlight_factor = 1, highlight_alpha = 0.5)
}
\arguments{
\item{comp_obj}{Object outputed by the function \code{\link{processTrendComparison}}.}

\item{ymax}{Numeric representing the maximum number of UMIs to show in the y axis of the trend plot. Default: Maximum number of UMIs.}

\item{xlim}{Numeric vector with the upstream and downstream limits of the region to be plotted. Default: Maximum range in the trend object.}

\item{plot_domainogram}{Logical for plotting the domainogram (not implemented yet).}

\item{protein_coding}{Show only protein coding genes in the gene track. Default: TRUE.}

\item{legend}{Named character vector with the names to use for the legend and colors to use for plotting. The first
element should be the "Treatment" and the second one the sample used as reference.}

\item{highlight}{If the input is a gene or a list of gene names will highlight the coordinates of such genes.
If the input is a GRanges object, will highlight the coordinates represented there.}

\item{highlight_factor}{Expansion factor for the coordinates of the highlight. The default is 1, meaning that will plot
the coordinates/gene as given. When increased, will expand the area of the highlight.}

\item{highlight_alpha}{Numeric character from 0 to 1 representing the transparency of the highlight. Default: 0.5}
}
\value{
A ggplot2 object with the full comparative plot.
}
\description{
Using a comparison object returned from \code{\link{processTrendComparison}}, it draws a comparative plot of the
two profiles and a domainogram of the difference.
}
