% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4CObject.R
\name{plotDgram}
\alias{plotDgram}
\title{Plot single domainogram}
\usage{
plotDgram(umi4c_obj = umi4c_obj, xlim = NULL)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{xlim}{Numeric vector with the upstream and downstream limits of the region to be plotted. Default: Maximum range in the trend object.}
}
\description{
Plot single domainogram
}
