% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparisonUMI4C.R
\name{plotDgramComp}
\alias{plotDgramComp}
\title{Plot comparative domainogram}
\usage{
plotDgramComp(comp_obj, xlim = NULL, legend)
}
\arguments{
\item{comp_obj}{Object outputed by the function \code{\link{processTrendComparison}}.}

\item{xlim}{Numeric vector with the upstream and downstream limits of the region to be plotted. Default: Maximum range in the trend object.}

\item{legend}{Named character vector with the names to use for the legend and colors to use for plotting. The first
element should be the "Treatment" and the second one the sample used as reference.}
}
\description{
Plot comparative domainogram
}
