% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot genes}
\usage{
plotGenes(window, protein_coding = TRUE, longest = TRUE, xlim = NULL)
}
\arguments{
\item{window}{GRanges object with coordinates from which to plot existing genes.}

\item{protein_coding}{Plot only protein coding genes. Default: TRUE}
}
\description{
Plot genes
}
