% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callInteractions.R
\name{plotInteractions}
\alias{plotInteractions}
\title{Plot interactions}
\usage{
plotInteractions(gr_interactions, xlim = NULL, significant = TRUE)
}
\arguments{
\item{gr_interactions}{GRangesList outputed by \code{zscoreUMI4C}.}

\item{xlim}{Limits for the plot x axis (genomic coordinates).}

\item{significant}{Logical indicating whether to plot only significant interactions
(default: TRUE).}
}
\value{
Produces a ggplot2 plot showing the queried interactions at each analyzed
sample.
}
\description{
Plot interactions
}
