% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparisonUMI4C.R
\name{plotTrendComp}
\alias{plotTrendComp}
\title{Plot comparative trend}
\usage{
plotTrendComp(comp_obj, ymax = NULL, xlim = NULL, legend)
}
\arguments{
\item{comp_obj}{Object outputed by the function \code{\link{processTrendComparison}}.}

\item{ymax}{Numeric representing the maximum number of UMIs to show in the y axis of the trend plot. Default: Maximum number of UMIs.}

\item{xlim}{Numeric vector with the upstream and downstream limits of the region to be plotted. Default: Maximum range in the trend object.}

\item{legend}{Named character vector with the names to use for the legend and colors to use for plotting. The first
element should be the "Treatment" and the second one the sample used as reference.}
}
\description{
Plot comparative trend
}
