% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4CObject.R
\name{plotUMI4CObject}
\alias{plotUMI4CObject}
\title{Plot a single UMI-4C profile}
\usage{
plotUMI4CObject(umi4c_obj, ymax = NULL, xlim = NULL, plot_raw = T,
  plot_domainogram = T, protein_coding = T, highlight = NULL,
  highlight_factor = 1, highlight_alpha = 0.5)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ymax}{Numeric representing the maximum number of UMIs to show in the y axis of the trend plot. Default: Maximum number of UMIs.}

\item{xlim}{Numeric vector with the upstream and downstream limits of the region to be plotted. Default: Maximum range in the trend object.}

\item{plot_raw}{Logical indicating wheter to plot the raw UMI-4C counts as points.}

\item{plot_domainogram}{Logical for plotting the domainogram (not implemented yet).}

\item{protein_coding}{Show only protein coding genes in the gene track. Default: TRUE.}

\item{highlight}{If the input is a gene or a list of gene names will highlight the coordinates of such genes.
If the input is a GRanges object, will highlight the coordinates represented there.}

\item{highlight_factor}{Expansion factor for the coordinates of the highlight. The default is 1, meaning that will plot
the coordinates/gene as given. When increased, will expand the area of the highlight.}

\item{highlight_alpha}{Numeric character from 0 to 1 representing the transparency of the highlight. Default: 0.5}
}
\value{
A ggplot2 object with the full plot for the UMI-4C profile.
}
\description{
Using a UMI-4C profile object returned from \code{\link{processUMI4C}}, it draws a plot of the profile and its domainogram.
}
