% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{prep}
\alias{prep}
\title{Prep fastq files at a given barcode.}
\usage{
prep(fq_R1, fq_R2, bait_seq, bait_pad, res_enz, prep_dir)
}
\arguments{
\item{fq_R1}{Fastq file R1.}

\item{fq_R2}{Fastq file R2..}

\item{bait_seq}{Character containing the bait primer sequence.}

\item{bait_pad}{Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{prep_dir}{Prep directory.}
}
\description{
Prep fastq files at a given barcode.
}
