% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{prepR}
\alias{prepR}
\title{Prepare UMI4C data}
\usage{
prepR(
  fastq_dir,
  wk_dir,
  bait_seq,
  bait_pad,
  res_enz,
  path_venv,
  fastq_multx = "fastq-multx"
)
}
\arguments{
\item{fastq_dir}{Path of the directory containing the FastQ files (compressed or uncompressed).}

\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{bait_seq}{Character containing the bait primer sequence.}

\item{bait_pad}{Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{path_venv}{Path for the UMI4Cats virtual environment generated using the \code{\link{installVenv}}.}

\item{fastq_multx}{Path or character for calling the \code{fastq-multx} script to demultiplex reads.}
}
\description{
Prepare the FastQ files for the analysis by selecting reads with bait and
adding the respective UMI identifier for each read in its header.
}
\examples{
\dontrun{
prepR(fastq_dir="raw_fastq",
      wk_dir="SOCS1",
      bait_seq="CCCAAATCGCCCAGACCAG",
      bait_pad="GCGCG",
      res_enz="GATC",
      path_venv="~/venvs/UMI4Cats_venv/",
      fastq_multx="fastq-multx")
}

}
\seealso{
\code{\link{contactsUMI4C}}
}
