% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepScript}
\alias{prepScript}
\title{Umi4c preprocessing}
\usage{
prepScript(input, cores, bait_seq, bait_pad, re, ref_gen, wk_dir,
  fastqmultx = "fastq-multx")
}
\arguments{
\item{input}{Path with the fastq (compressed or descompressed)}

\item{cores}{Number of cores to use in the analysis}

\item{bait_seq}{Bait sequence.}

\item{re}{Restriction enzyme sequence.}

\item{ref_gen}{Reference genome in fasta format.}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis.}

\item{fastqmultx}{Fastq-multx pàth}
}
\description{
This function filters the reads keeping the one that present bait sequence, pad and restriction enzyme. 
Moreover, it introduces the umi sequence in the header of the fasta files
}
