% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTrends.R
\name{processTrendComparison}
\alias{processTrendComparison}
\title{Compare two UMI-4C profiles}
\usage{
processTrendComparison(umi4c_obj, ref_umi4c_obj, min_win_cov = NULL,
  norm_bins = 10^(3:6), factor = 0.025, sd = 2)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ref_umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile to use as reference.
\code{umi4c_obj} will be scaled to the \code{ref_umi4c_obj}.}

\item{min_win_cov}{Minimum number of UMIs requiered in a region to perform the adaptaive smoothing. Default: NULL,
number is calculated by the function \code{\link{getMinWinMols}} using the factor specified below.}

\item{norm_bins}{Bins used to normalize the number of UMIs from \code{umi4c_obj} to \code{umi4c_ref}. Default: 10^(3:6).}

\item{factor}{Factor representing the proportion of UMIs from the total UMIs in the analyzed region needed as a minimum
to perform the adaptative smoothing.}

\item{sd}{Standard devition value for the adaptative smoothing. Default: 2 (extracted from Schwartzman et al).}
}
\value{
A list object with the following named objects:
}
\description{
Scales the umi4c_obj to the ref_umi4c_obj, creating all necessary elements to produce the comparative plot.
}
