% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{smoothTrendAdaptative}
\alias{smoothTrendAdaptative}
\title{Adaptative smoothing of the UMI-4C trend}
\usage{
smoothTrendAdaptative(UMI4C, sd = 2)
}
\arguments{
\item{UMI4C}{UMI4C objected created with the \code{\link{UMI4C}} function.}

\item{sd}{Standard devition value for the adaptative smoothing. Default: 2 (extracted from Schwartzman et al).}

\item{dgram}{Output of the function \code{\link{createDomainogram}}.}
}
\value{
A data.frame containing the necessary information to draw the different elements of the UMI-4C trend:
\itemize{
\item \strong{coord}: Coordinates of the fragment end.
\item \strong{trend}: Smoothed UMIs value according to min_win_cov.
\item \strong{scale}: Scale used (# of fragends to merge) to achieve min_win_cov.
\item \strong{group}: Fragment ends grouped according to postion using bait as reference (upstream or downstream).
\item \strong{devP}: Upper bound of standard deviation.
\item \strong{devM}: Lower bound of standard deviation.
}
}
\description{
Performs adaptative smoothing to plot a UMI-4C contact trend. This means that joins fragment ends together until the
minimum number of UMIs in the joined fragment is >= min_win_cov.
}
