% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTrends.R
\name{smoothTrendAdaptativeComp}
\alias{smoothTrendAdaptativeComp}
\title{Adaptative smoothing of scaled trend}
\usage{
smoothTrendAdaptativeComp(umi4c_obj, ref_umi4c_obj, min_win_cov = 50,
  sd = 2)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ref_umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile to use as reference.
\code{umi4c_obj} will be scaled to the \code{ref_umi4c_obj}.}

\item{min_win_cov}{Minimum number of UMIs requiered in a region to perform the adaptaive smoothing. Default: NULL,
number is calculated by the function \code{\link{getMinWinMols}} using the factor specified below.}

\item{sd}{Standard devition value for the adaptative smoothing. Default: 2 (extracted from Schwartzman et al).}
}
\description{
Will perform adaptative smoothing will scaling one profile to the reference UMI-4C profile.
}
