% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{splitUMI4C}
\alias{splitUMI4C}
\title{Split UMI4C reads}
\usage{
splitUMI4C(wk_dir, prep_dir = "", cut_seq_5p, cut_seq_3p)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{cut_seq_5p}{5' restriction enzyme cut sequence. For example, for DpnII |GATC \code{cut_seq_5p=""}.}

\item{cut_seq_3p}{3' restriction enzyme cut sequence. For example, for DpnII |GATC \code{cut_seq_3p="GATC"}.}
}
\description{
Split the prepared reads using the restrition enzyme information.
}
\examples{
\dontrun{
splitUMI4C(wk_dir="SOCS1",
       cut_seq_5p="",
       cut_seq_3p="GATC")
}
}
