% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareTrends.R
\name{summarizeRaw}
\alias{summarizeRaw}
\title{Summarize raw UMIs}
\usage{
summarizeRaw(umi4c_obj, ref_umi4c_obj)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ref_umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile to use as reference.
\code{umi4c_obj} will be scaled to the \code{ref_umi4c_obj}.}
}
\description{
Summarize raw UMIs
}
