% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{theme}
\alias{theme}
\title{Theme}
\usage{
theme(...)
}
\arguments{
\item{...}{Additional arguments to pass to the theme call from ggplot2.}
}
\value{
ggplot2 theme.
}
\description{
Theme
}
\examples{
library(ggplot2)

ggplot(
    iris,
    aes(Sepal.Length, Sepal.Width)
) +
    geom_point() +
    theme()
}
