% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{themeXYblank}
\alias{themeXYblank}
\title{Theme Y blank}
\usage{
themeXYblank(...)
}
\arguments{
\item{...}{Additional arguments to pass to the theme call from ggplot2.}
}
\value{
ggplot2 theme with a blank X and Y axis.
}
\description{
Theme Y blank
}
\examples{
library(ggplot2)

ggplot(iris,
       aes(Sepal.Length, Sepal.Width)) +
  geom_point() +
  themeXYblank()
}
