% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{themeYblank}
\alias{themeYblank}
\title{Theme Y blank}
\usage{
themeYblank(...)
}
\arguments{
\item{...}{Additional arguments to pass to the theme call from ggplot2.}
}
\value{
ggplot2 theme with a blank Y axis.
}
\description{
Theme Y blank
}
