% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R, R/docsMethods.R
\docType{methods}
\name{dgram}
\alias{dgram}
\alias{dgram<-}
\alias{bait}
\alias{trend}
\alias{results}
\alias{dgram,UMI4C-method}
\alias{dgram<-,UMI4C-method}
\alias{bait,UMI4C-method}
\alias{trend,UMI4C-method}
\alias{results,UMI4C-method}
\alias{UMI4C-methods}
\title{UMI4C class methods}
\usage{
dgram(object)

dgram(object) <- value

bait(object)

trend(object)

results(object, format = "GRanges", counts = TRUE)

\S4method{dgram}{UMI4C}(object)

\S4method{dgram}{UMI4C}(object) <- value

\S4method{bait}{UMI4C}(object)

\S4method{trend}{UMI4C}(object)

\S4method{results}{UMI4C}(object, format = "GRanges", counts = FALSE)
}
\arguments{
\item{object}{a \code{UMI4C-class} object.}

\item{value}{Alternative list of dgrams to replace the current slot.}

\item{format}{Either "GRanges" (default) or "data.frame", indicating the format output of the results.}

\item{counts}{Logical indicating whether counts for the different region should be provided. Default: FALSE.}
}
\value{
There are several accessors to easily retrive information from a \code{UMI4C-class} object:
\itemize{
  \item \code{dgram}: Returns a named list with the output domainograms for each sample.
  \item \code{bait}: Returns a \linkS4class{GRanges} object with the position of the bait.
  \item \code{trend}: Returns a data.frame in long format with the values of the adapative smoothen trend.
  \item \code{results}: Returns a \linkS4class{GRanges} or data.frame with the results of the differential analysis.
}
}
\description{
This page contains a summary of the different methods used to access the information contained
inside the \code{UMI4C-class} object. See the details section for more information on the different accessors.
}
\examples{
\dontrun{
# Create example UMI4C-class object with helper function
umi <- makeUMI4Cexample()

# Access the different information inside the UMI4C object
dgram(umi)
bait(umi)
trend(umi)

# Perform differential test
umi <- fisherUMI4C(umi)
results(umi)
}
}
\seealso{
UMI4C, UMI4C-class
}
