% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FourC}
\alias{FourC}
\alias{FourC-class}
\alias{UMI4C}
\title{UMI4C-class}
\usage{
UMI4C(colData, viewpoint_name)
}
\arguments{
\item{colData}{Data.frame object containing all the sample information required for the analysis:
\enumerate{
\item sampleID: Name for the sample.
\item replicate: Replicate number.
\item condition: Experimental condition or tissue.
\item file: Path for the output file containing UMI contact information.
}}

\item{metadata}{List containing the following elements:
\enumerate{
\item bait: GRanges object representing the position of the bait used for the analysis.
}}

\item{assayData}{Matrix containing the ID for the restriction fragment as row.names and each column representing
the number of raw UMIs for a specific sample.}
}
\description{
UMI4C-class
}
\note{
The \code{UMI4C} object extends the \code{DESeqDataSet} class.
}
\keyword{package}
