% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMI4Cats.R
\docType{package}
\name{UMI4Cats}
\alias{UMI4Cats}
\title{UMI4Cats: A package for analyzing UMI-4C chromatin contact data}
\description{
The UMI4Cats package provides functions for the pre-processing, analysis and
visualization of UMI-4C chromatin contact data.
}
\section{Pre-processing}{

The pre-processing functions ...
}

\section{Analysis}{

The analysis functions ...
}

\section{Visualization}{

The plotting functions ...
}

