% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi4CatsContacts.R
\name{umi4CatsContacts}
\alias{umi4CatsContacts}
\title{Umi4c analysis}
\usage{
alignmentR(raw_dir, wk_dir, bait_seq, bait_pad, res_e, fastqmultx)
}
\arguments{
\item{raw_dir}{Path with the fastq (compressed or descompressed)}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{threads}{Number of threads to use in the analysis}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{res_e}{Restriction enzyme sequence}

\item{ref_gen}{Reference genome in fasta format}

\item{genomic_track}{Genomic track path}

\item{trimmomatic}{Trimmomatic path. By default trimmomatic}

\item{bowtie2}{Bowtie2 path. By default bowtie2}

\item{fastqmultx}{Fastq-multx pàth. By default fastq-multx}

\item{samtools}{Samtools path. By default samtools}

\item{cut_pos}{Nucleotide where restriction enzyme cuts (zero-base)}

\item{genomic_track}{Genomic track path generated by the generateGenomicTrack function of UMI4Cats package}
}
\description{
This function parses, clean, splits, aligns and analyses fastq files counting chromatin contacts between the viewpoint and 
the surrounding loci. The output is a tsv file used for the generation of 4C profile.
}
\examples{
raw_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/raw'
wk_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/MLH1_ctrl_umi4cats_python'
threads = 1
bait_seq = 'ACCTAGAAGGATATGCGCTTGC'
bait_pad = 'GCGTTAGA'
res_e = "GATC"
cut_pos = "0" # should be a string for properly working
fastqmultx = 'fastq-multx'
bowtie2 = 'bowtie2'
ref_gen = '/imppc/labs/lplab/share/marc/refgen/hg19/hg19.fa'
samtools = 'samtools'
genomic_track = '/imppc/labs/lplab/share/marc/epimutations/processed/genomicTracks/genomic_tracks_hg19/dpnII_genomicTrack'
py_umi4cats = '/imppc/labs/lplab/share/marc/repos/umi4cats/python/umi4catsBuilder.py'
umi4CatsContacts(py_umi4cats,
                 raw_dir,
                 wk_dir,
                 threads,
                 bait_seq,
                 bait_pad,
                 res_e,
                 ref_gen,
                 genomic_track,
                 trimmomatic,
                 bowtie2,
                 fastqmultx,
                 samtools)
                 
}
