% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi4catsVenvInst.R
\name{umi4catsVenvInst}
\alias{umi4catsVenvInst}
\title{umi4Cats virtualenv}
\usage{
umi4catsVenvInst()
}
\description{
Install automatically virtualenv from requirements.txt
}
\examples{
umi4catsVenvInst()
}
