% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{umiCount}
\alias{umiCount}
\title{Counts umis for a given bam file.}
\usage{
umiCount(
  filtered_bam_R1,
  filtered_bam_R2,
  digested_genome_gr,
  pos_viewpoint,
  res_enz,
  count_dir,
  filter_bp = 1e+07
)
}
\arguments{
\item{filtered_bam_R1}{R1 bam file,}

\item{filtered_bam_R2}{R2 bam file,}

\item{pos_viewpoint}{Vector consist of chromosome, start and end position of the viewpoint.}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{count_dir}{Counter directory.}

\item{bowtie_index}{Bowtie index file for the reference genome.}
}
\description{
Counts umis for a given bam file.
}
