% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiCounter.R
\name{umiCounter}
\alias{umiCounter}
\title{Umi4c umiCount}
\usage{
umiCounter(wk_dir, cores, bait_seq, bait_pad, re, ref_gen, genomic_track,
  bowtie2 = "bowtie2")
}
\arguments{
\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{ref_gen}{Reference genome in fasta format}

\item{genomic_track}{Genomic track path}

\item{bowtie2}{Bowtie2 path. By default bowtie2}
}
\description{
This function relates the segments of reads generated in the split phase with the fragments of a defined genome track 
(which can be generated by genomicTrackGenerator.R), and counts the number of UMIs for every contact.
}
