% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiCounterR.R
\name{umiCounterR}
\alias{umiCounterR}
\title{umi4Cats umiCounter}
\usage{
umiCounterR(py_umi4cats, raw_dir, wk_dir, bait_seq, bait_pad, res_e, fastqmultx)
}
\arguments{
\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{bowtie2}{Bowtie2 path. By default bowtie2}

\item{ref_gen}{Reference genome in fasta format}

\item{samtools}{Samtools path. By default samtools}

\item{genomic_track}{Genomic track path}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{res_e}{Restriction enzyme sequence}
}
\description{
Count number of UMIs for every ligation taking ligation position and UMI mismatches into acount. It generates a tsv file 
for creating the 4c profile.
}
