% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialNbinomWaldTestUMI4C.R
\name{vstUMI4C}
\alias{vstUMI4C}
\title{Variance stabilizing transformation
Using a DDS object performs a variance stabilizing transformation from DESeq2 package
to the UMI4C counts}
\usage{
vstUMI4C(dds)
}
\arguments{
\item{dds}{DDS object generated by \code{UMI4C2dds}}
}
\value{
DDS object with variance stabilizing transformation counts
}
\description{
Variance stabilizing transformation
Using a DDS object performs a variance stabilizing transformation from DESeq2 package
to the UMI4C counts
}
\details{
This function estimate the size factors and dispersions of the counts
base on \code{\link[DESeq2]{DESeq}} for infering the VST distribution and
transform raw UMI4C counts.
}
