% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentR.R
\name{alignmentR}
\alias{alignmentR}
\title{umi4Cats alignment}
\usage{
alignmentR(py_umi4cats, raw_dir, wk_dir, bait_seq, bait_pad, res_e, fastqmultx)
}
\arguments{
\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{threads}{Number of threads to use in the analysis}

\item{bowtie2}{Bowtie2 path. By default bowtie2}

\item{ref_gen}{Reference genome in fasta format}

\item{samtools}{Samtools path. By default samtools}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{res_e}{Restriction enzyme sequence}
}
\description{
Align splitted UMI reads agains a pre-defined reference genome using Bowtie2
}
