% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{alignmentR}
\alias{alignmentR}
\title{UMI4C alignment}
\usage{
alignmentR(
  wk_dir,
  bait_seq,
  bait_pad,
  res_enz,
  ref_gen,
  threads = 1,
  path_venv
)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{bait_seq}{Character containing the bait primer sequence.}

\item{bait_pad}{Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{ref_gen}{Path for the reference genome to use for the alignment (fasta format).}

\item{threads}{Number of threads to use in the analysis.}

\item{path_venv}{Path for the UMI4Cats virtual environment generated using the \code{\link{installVenv}}.}
}
\description{
Align splitted UMI-4C reads to a reference genome using Bowtie2.
}
\examples{
\dontrun{
alignmentR(wk_dir="SOCS1",
           bait_seq="CCCAAATCGCCCAGACCAG",
           bait_pad="GCGCG",
           res_enz="GATC",
           ref_gen="~/data/reference_genomes/hg19/hg19.fa",
           threads=1,
           path_venv="~/venvs/UMI4Cats_venv/")
}

}
