% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{alignmentUMI4C}
\alias{alignmentUMI4C}
\title{UMI4C alignment}
\usage{
alignmentUMI4C(wk_dir, pos_viewpoint, ref_gen, threads = 1, rm_tmp = TRUE)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{ref_gen}{Path for the reference genome to use for the alignment (fasta format).}

\item{threads}{Number of threads to use in the analysis.}

\item{rm_tmp}{Logical indicating whether to remove temporary files (sam and intermediate bams) (TRUE or FALSE, defaults to TRUE).}
}
\description{
Align splitted UMI-4C reads to a reference genome using Bowtie2.
}
\examples{
\dontrun{
alignmentR(wk_dir="SOCS1",
           bait_seq="CCCAAATCGCCCAGACCAG",
           bait_pad="GCGCG",
           res_enz="GATC",
           ref_gen="~/data/reference_genomes/hg19/hg19.fa")
}

}
