% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAdaptativeTrend.R
\name{calculateAdaptativeTrend}
\alias{calculateAdaptativeTrend}
\title{Adaptative smoothing of normalized trend}
\usage{
calculateAdaptativeTrend(umi4c, sd = 2, normalized = FALSE)
}
\arguments{
\item{umi4c}{umi4c object.}

\item{sd}{Standard deviation}

\item{normalized}{Logical indicating whether smoothen trends should be normalized to the sample with less UMIs.}
}
\description{
Will perform adaptative smoothing will scaling one profile to the reference UMI-4C profile.
}
