% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeUMI4C.R
\name{calculateAdaptativeTrend}
\alias{calculateAdaptativeTrend}
\title{Adaptative smoothing of normalized trend}
\usage{
calculateAdaptativeTrend(umi4c, sd = 2, normalized = TRUE)
}
\arguments{
\item{umi4c}{\linkS4class{UMI4C} object as generated by \code{\link{makeUMI4C}}.}

\item{sd}{Stantard deviation for adaptative trend.}

\item{normalized}{Logical indicating whether UMI-4C profiles should be
normalized to the \code{ref_umi4c} sample/group. Default: TRUE}
}
\value{
Calculates the adaptative trend considering the minimum number of
molecules to use for merging different restriction fragments. It also
calculates the geometric mean of the coordinates of the merged restriction
fragments.
}
\description{
Will perform adaptative smoothing will scaling one profile to the reference
UMI-4C profile.
}
