% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{calculateAdaptativeTrend}
\alias{calculateAdaptativeTrend}
\title{Adaptative smoothing of normalized trend}
\usage{
calculateAdaptativeTrend(umi4c, sd = 2, normalized = TRUE)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{sd}{Stantard deviation for adaptative trend.}

\item{normalized}{Logical indicating whether UMI4C profiles should be normalized to the sample with less UMIs (ref_umi4c). Default: TRUE}
}
\description{
Will perform adaptative smoothing will scaling one profile to the reference UMI-4C profile.
}
