% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{calculateDomainogram}
\alias{calculateDomainogram}
\title{Create Domainogram}
\usage{
calculateDomainogram(umi4c, scales = 5:150, normalized = TRUE)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{scales}{Integer vector indicating the number of scales to use for the domainogram creation. Default: 5:150.}

\item{normalized}{Logical whether the the resulting domainograms should be normalized or not. Default: TRUE.}
}
\value{
A matrix where the first column represents the fragment end coordinates (start) and the rest represent the
number of UMIs found when using a specific scale.
}
\description{
Using as input the raw UMIs, this package creates a domainogram for the supplied scales.
}
