% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDomainogram.R
\name{createDomainogram}
\alias{createDomainogram}
\title{Create Domainogram}
\usage{
createDomainogram(input, bait_coordinates, scales = 5:150,
  bait_exclusion = 3000, bait_upstream = 5e+05,
  bait_downstream = 5e+05)
}
\arguments{
\item{input}{A two-column data.frame where the first column represents the fragend and the second the number of UMIs.}

\item{scales}{Scales to use for smoothing and plotting the domainogram. Default: 1:150.}

\item{bait_exclusion}{Region around the bait (in bp) to be excluded from the analysis. Default: 3kb.}

\item{bait_upstream}{Number of bp upstream of the bait to use for the analysis. Default: 500kb.}

\item{bait_downstream}{Number of bp downstream of the bait to use for the analysis. Default: 500kb.}

\item{bait_coordintes}{GRanges object with the coordinates of the bait used for the experiment.}
}
\description{
Using as input the raw UMIs, this package creates a domainogram for the supplied scales.
}
