% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{createDomainogram}
\alias{createDomainogram}
\title{Create Domainogram}
\usage{
createDomainogram(input, bait_coordinates, scales = 5:150,
  bait_exclusion = 3000, bait_upstream = 5e+05,
  bait_downstream = 5e+05)
}
\arguments{
\item{input}{A five-column data.frame or a character string indicating its path, where the first two columns represent the bait position ("chr" and "pos), and the following three
columns represent the contact position and the number of UMIs supporting that contact ("chr", "pos" and "UMIs").}

\item{bait_coordinates}{GRanges object with the coordinates of the bait used for the experiment. If left empty, will take coordinates from the input (first two columns).}

\item{scales}{Scales to use for smoothing and plotting the domainogram. Scales indicate how many fragment
ends are merged together to obtain the normalized number of UMIs. Default: 5:150.}

\item{bait_exclusion}{Region around the bait (in bp) to be excluded from the analysis. Default: 3kb.}

\item{bait_upstream}{Number of bp upstream of the bait to use for the analysis. Default: 500kb.}

\item{bait_downstream}{Number of bp downstream of the bait to use for the analysis. Default: 500kb.}
}
\value{
A matrix where the first column represents the fragment end coordinates (start) and the rest represent the
number of UMIs found when using a specific scale.
}
\description{
Using as input the raw UMIs, this package creates a domainogram for the supplied scales.
}
