% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsUMI4C.R
\name{createStatsTable}
\alias{createStatsTable}
\title{Create stats table}
\usage{
createStatsTable(wk_dir)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the
UMI-4c analysis.}
}
\value{
Returns a data.frame summarizing all the different statistics
for each sample analyzed in \code{wk_dir}.
}
\description{
Create a statistical summary of the UMI-4C experiments analyzed with
\code{contactsUMI4C}.
}
