% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialUMI4C.R
\name{deseq2UMI4C}
\alias{deseq2UMI4C}
\title{Differential UMI4C contacts using DESeq2 Wald Test}
\usage{
deseq2UMI4C(
  umi4c,
  design = ~condition,
  normalized = TRUE,
  padj_method = "fdr",
  query_regions = NULL,
  padj_threshold = 0.05,
  ...
)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the
\code{UMI4C} constructor.}

\item{design}{A \code{formula} or \code{matrix}. The formula expresses how
the counts for each fragment end depend on the variables in \code{colData}.
See  \code{\link[DESeq2]{DESeqDataSet}}}

\item{normalized}{Logical indicating if the function shoult return normalized
or raw UMI counts.}

\item{padj_method}{The method to use for adjusting p-values, see
\code{\link[stats]{p.adjust}}}

\item{query_regions}{\code{GRanges} object or \code{data.frame} containing
the coordinates of the genomic regions you want to use to perform the
analysis in specific genomic intervals.}

\item{padj_threshold}{Numeric indicating the adjusted p-value threshold to
use to define significant differential contacts.}

\item{...}{Other arguments to be passed to \code{\link[DESeq2]{DESeq}}
function.}
}
\value{
Differential results using DESeq2.
}
\description{
Differential UMI4C contacts using DESeq2 Wald Test
}
