% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialContacts.R
\name{differentialContacts}
\alias{differentialContacts}
\title{Test differential contacts}
\usage{
differentialContacts(umi4c_obj, ref_umi4c_obj, regions,
  window_total = 1e+06, padj_method = "fdr")
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ref_umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile to use as reference.
\code{umi4c_obj} will be scaled to the \code{ref_umi4c_obj}.}

\item{window_total}{Region around the bait to use for getting total number of UMIs (the region around the bait
is already removed when running processUMI4()). Default: 2Mb.}

\item{padj_method}{Method to use for adjusting the p-values for multiple testing. See p.adjust().}
}
\description{
Performs fisher's exact tests for each region supplied comparing the umi4c_obj to the reference umi4c_obj,
thus testing for differential contacts in such regions.
}
