% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestGenome.R
\name{digestGenome}
\alias{digestGenome}
\title{Digest reference genome}
\usage{
digestGenome(res_enz, cut_pos, name_RE, ref_gen, out_path = "digested_genome/")
}
\arguments{
\item{res_enz}{Character containing the restriction enzyme sequence.
#' @param cut_pos Numeric indicating the nucleotide position where restriction enzyme cuts (zero-based) (for example, for DpnII is 0).}

\item{name_RE}{Restriction enzyme name .}

\item{ref_gen}{A BSgenome object of the reference genome.}

\item{out_path}{Output path where to save the genomic track. The default is a directory named \code{digested_genome/}
created in your working directory.}
}
\value{
Creates a tab-delimited file in \code{out_path} named "\code{ref_gen_name_RE.tsv}", containing the
coordinates for all fragments generated by digesting the genome using the given cutting sequences.
}
\description{
Performs an \emph{in silico} digestion of a given reference genome using a given restriction enzyme sequence.
}
\examples{
\dontrun{
library(BSgenome.Hsapiens.UCSC.hg19)
ref_gen <- BSgenome.Hsapiens.UCSC.hg19

hg19_dpnii <- digestGenome(res_enz="GATC",
                           cut_pos=0
                           name_RE="DpnII",
                           ref_gen=ref_gen,
                           out_path="digested_genome/")
}
}
