% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUMI4Cexample.R
\name{downloadUMI4CexampleData}
\alias{downloadUMI4CexampleData}
\title{Download UMI4Cats example datasets}
\usage{
downloadUMI4CexampleData(
  output_dir = "./",
  file_dir = "http://gattaca.imppc.org/genome_browser/lplab/UMI4Cats_data.tar.gz",
  use_sample = FALSE,
  only_index = FALSE
)
}
\arguments{
\item{output_dir}{Output directory for the datasets.}

\item{file_dir}{Path to the compressed IRB database file.}

\item{use_sample}{Logical value for use a smaller sample of the examples.}

\item{only_index}{Logical indicating if only reduced bowtie2 index for
chr16 should be downloaded.}
}
\value{
It creates the \code{output_dir} with the example UMI-4C files used
 by the vignette.
}
\description{
Downloads the required UMI4Cats example datasets.
}
\examples{
## Use subsample of data to make example faster.
## Remove `use_sample=TRUE` or set to `FALSE` to
## download full dataset

path <- downloadUMI4CexampleData(use_sample = TRUE)

unlink(path, recursive =TRUE )
}
