% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUMI4Cexample.R
\name{downloadUMI4CexampleData}
\alias{downloadUMI4CexampleData}
\title{Download UMI4Cats example datasets}
\usage{
downloadUMI4CexampleData(
  output_dir = "./",
  file_dir = "http://gattaca.imppc.org/genome_browser/lplab/UMI4Cats_data.tar.gz"
)
}
\arguments{
\item{output_dir}{Output directory for the datasets.}

\item{file_dir}{Path to the compressed IRB database file.}
}
\description{
Downloads the required UMI4Cats example datasets.
}
