% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialUMI4C.R
\name{fisherUMI4C}
\alias{fisherUMI4C}
\title{Differential UMI4C contacts using Fisher's Exact test}
\usage{
fisherUMI4C(umi4c, grouping = "condition", query_regions,
  resize = NULL, window_size = 5000, filter_low = 50,
  padj_method = "fdr", padj_threshold = 0.05)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{grouping}{Name of the \code{colData(UMI4C)} column that will be used for grouping the UMI4C profiles
and performing differential analyses.}

\item{query_regions}{\code{GenomicRanges} object containing the coordinates for the regions where to
perform the differential analysis.}

\item{resize}{Width in base pairs for resizing the query regions. Default: no resizing.}

\item{window_size}{If \code{query_regions} are not defined, wil bin region in \code{window_size} bp and
perform the analysis using this windows.}

\item{filter_low}{Either the minimum median UMIs requiered to perform Fisher's Exact test or \code{FALSE} for
performing the test in all windows.}

\item{padj_method}{Method for adjusting p-values. See \code{\link[stats]{p.adjust}} for the different methods.}

\item{padj_threshold}{Numeric indicating the adjusted p-value threshold to use to define significant differential
contacts.}
}
\description{
Using the UMIs inside \code{query_regions} performs Fisher's Exact test to calculate significant differences
between contact intensities.
}
\details{
In order to calculate differential contact intensities, this fuction first sums all UMIs for each
grouping condition. Then, calculates the overlap of fragment ends with \code{query_regions}. The resulting number
of UMIs in each \code{query_region} will be the sum of UMIs in all overlapping fragment ends. Then, a contingency
table for each \code{query_reegions} is created as follows:
\tabular{rcc}{
     \tab \emph{query_region} \tab \emph{region}\cr
    \emph{Reference} \tab n1 \tab N1-n1\cr
    \emph{Condition} \tab n2 \tab N2-n2
    }
and the Fisher's Exact test is performed.
}
