% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicTrackGenerator.R
\name{genomicTrackGenerator}
\alias{genomicTrackGenerator}
\title{Genomic track generator}
\usage{
genomicTrackGenerator(pos, re, nameRe, refgen, output)
}
\arguments{
\item{pos}{Restiction enzyme cut. For example pos 5 in HgaI GACGC indicates cleavage as follows: 5' GACGC^NNNNN.}

\item{re}{Restriction enzyme sequence.}

\item{nameRe}{Name restriction enzyme.}

\item{refgen}{Reference genome used for the generation of the genomic tracks.}

\item{output}{Output path where to save the genomic tracks.}
}
\description{
This function filters the reads keeping the one that present bait sequence, pad and restriction enzyme. 
Moreover, it introduces the umi sequence in the header of the fasta files
}
