% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicTrackGenerator.R
\name{genomicTrackGenerator}
\alias{genomicTrackGenerator}
\title{Genomic track generator}
\usage{
genomicTrackGenerator <- function(cp5p, cs3p, nameRe, refgen, output)
}
\arguments{
\item{cs5p}{5' restriction enzyme cut sequence. In the GA|TC case "GA".}

\item{cs3p}{3' restriction enzyme cut sequence. In the GA|TC case "TC"}

\item{nameRe}{Restriction enzyme name .}

\item{refgen}{A BSgenome object of the reference genome.}

\item{outPath}{Output path where to save the genomic track.}
}
\description{
In silico digestion of a reference genome
}
\examples{
\dontrun{
output <- '/imppc/labs/lplab/share/marc/refgen/genomicTracksR'
the selected RE in this case is DpnII (|GATC), so the cs5p is "" and cs3p is GATC
cs5p <- ""
cs3p <- "GATC"
nameRe <- 'dpnII'
refgen <- BSgenome.Hsapiens.UCSC.hg19

genomicTrackGenerator <- function(cp5p, cs3p, nameRe, refgen, output)
}

}
