% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{getColors}
\alias{getColors}
\title{Get default colors}
\usage{
getColors(factors)
}
\arguments{
\item{factors}{Name of the factors that will be used for grouping variables.}
}
\value{
Depending on the number of factors it creates different color
palettes.
}
\description{
Get default colors
}
