% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{getNormalizationMatrix}
\alias{getNormalizationMatrix}
\title{Get normalization matrix}
\usage{
getNormalizationMatrix(
  umi4c,
  norm_bins = 10^(3:6),
  post_smooth_win = 50,
  r_expand = 1.2
)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{norm_bins}{Numeric vector with the genomic bins to use for normalization. Default: 1K, 10K, 100K, 1Mb.}

\item{post_smooth_win}{Numeric indicating the smoothing window to use. Default: 50.}

\item{r_expand}{Numeric indicanting the expansion value for normalization. Default: 1.2.}
}
\description{
Will return a normalization matrix.
}
\details{
bla
}
