% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installVenv.R
\name{installVenv}
\alias{installVenv}
\title{Install UMI4Cats python virtual environment}
\usage{
installVenv(path_venv = "venv_UMI4Cats")
}
\arguments{
\item{path_venv}{Path where to install the virtual environment.}
}
\value{
Creates a folder in \code{path_venv} containing all necessary python dependencies and packages to then run
UMI4Cats functions.
}
\description{
Automatically install python virtual environment and all necessary packages from \code{python/requirements.txt}.
}
\details{
Automatically creates a python virtual environment inside the defined \code{path_env}. User needs to have
\link[=https://www.python.org/download/releases/3.0/]{Python 3} installed and added to your path as \code{python 3}.
By default the path is created in the working directory and named \code{venv_UMI4Cats}.
}
\examples{
\dontrun{
 installVenv('~/venvs/UMI4Cats/')
}
}
