% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeUMICounter.R
\name{mergeUMICounter}
\alias{mergeUMICounter}
\title{umi4Cats mergeUMICounter}
\usage{
mergeUMICounter(genomic_track, wk_dir)
}
\arguments{
\item{genomic_track}{Genomic track path generated by the generateGenomicTrack function of UMI4Cats package}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}
}
\description{
Merge UMI4c counter with genome track positions.
}
\examples{
\dontrun{
wk_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/MLH1_ctrl_umi4cats_python'
genomic_track = '/imppc/labs/lplab/share/marc/epimutations/processed/genomicTracks/genomic_tracks_hg19/dpnII_genomicTrack'

mergeUMICounter(genomic_track, wk_dir)
}

}
