% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialNbinomWaldTestUMI4C.R
\name{nbinomWaldTestUMI4C}
\alias{nbinomWaldTestUMI4C}
\title{Differential UMI4C contacts using DESeq2 Wald Test
Takes the smooth monotone fit count values and infers the differential UMI4C 
contacts using DESeq2 Wald Test from 
 \code{\link[DESeq2]} package.}
\usage{
nbinomWaldTestUMI4C(dds)
}
\arguments{
\item{dds}{DDS object as generated by \code{smoothMonotoneUMI4C} with the smooth monotone fit counts}
}
\value{
DDS object with the DESeq2 Wald Test results, 
with results columns accessible with the \code{results} function.
}
\description{
Differential UMI4C contacts using DESeq2 Wald Test
Takes the smooth monotone fit count values and infers the differential UMI4C 
contacts using DESeq2 Wald Test from 
 \code{\link[DESeq2]} package.
}
\details{
This function back-transform fitted values to the
scale of raw counts and scale them across the samples to
obtin normalizaed factors unsing \code{normalizationFactors} from \code{DESeq2} package.
To detect differences between conditions, the DESeq2 Wald Test is applied to fitted counts with 
the normalization factors.
}
