% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotDifferential}
\alias{plotDifferential}
\title{Plot differential contacts}
\usage{
plotDifferential(umi4c, grouping = "condition", colors = NULL,
  xlim = NULL)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{grouping}{Name of the \code{colData(UMI4C)} column that will be used for grouping the profiles. If it is
vector of column names, will use both columns for grouping the samples.}

\item{colors}{Vector of colors to use for plotting for each group.}

\item{xlim}{Limits for the plot x axis (genomic coordinates).}
}
\value{
Produces a plot of the fold changes at the differential regions analyzed contained in
the UMI4C object.
}
\description{
Plot differential contacts
}
\examples{
umi <- makeUMI4Cexample()
umi_dif <- fisherUMI4C(umi, filter_low=30)
plotDifferential(umi_dif)

}
