% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotDomainogram}
\alias{plotDomainogram}
\title{Plot domainogram}
\usage{
plotDomainogram(umi4c, grouping = "condition",
  dgram_function = "quotient", colors = NULL, xlim = NULL)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{grouping}{Name of the \code{colData(UMI4C)} column that will be used for grouping the profiles. If it is
vector of column names, will use both columns for grouping the samples.}

\item{dgram_function}{The name of the function used for calculating the fold-change in the domainogram plot,
either "difference" or "quotient".}

\item{colors}{Vector of colors to use for plotting for each group.}

\item{xlim}{Limits for the plot x axis (genomic coordinates).}
}
\description{
Plot domainogram
}
