% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot genes}
\usage{
plotGenes(
  window,
  TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene,
  longest = TRUE,
  xlim = NULL
)
}
\arguments{
\item{window}{GRanges object with coordinates from which to plot existing genes.}

\item{TxDb}{TxDb object to use for drawing the genomic annotation.}

\item{longest}{Logical indicating whether to plot only the longest transcripts for each gene in the
gene annotation part.}

\item{xlim}{Limits for the plot x axis (genomic coordinates).}
}
\value{
Produces a plot with the genes found in the provided \code{window}.
}
\description{
Plot genes
}
\examples{
window <- GRanges("chr16:11298262-11400036")
plotGenes(window=window,
          TxDb=TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene)
}
