% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotGenes}
\alias{plotGenes}
\title{Plot genes}
\usage{
plotGenes(window, protein_coding = TRUE, longest = TRUE, xlim = NULL)
}
\arguments{
\item{window}{GRanges object with coordinates from which to plot existing genes.}

\item{protein_coding}{Logical indicating whether to only plot protein coding genes in the gene
annotation part.}

\item{longest}{Logical indicating whether to plot only the longest transcripts for each gene in the
gene annotation part.}

\item{xlim}{Limits for the plot x axis (genomic coordinates).}
}
\description{
Plot genes
}
