% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot single trend}
\usage{
plotTrend(umi4c_obj = umi4c_obj, ymax = NULL, xlim = NULL,
  plot_raw = T)
}
\arguments{
\item{umi4c_obj}{Object outputed by the function \code{\link{processUMI4C}} representing a single UMI-4C processed profile.}

\item{ymax}{Numeric representing the maximum number of UMIs to show in the y axis of the trend plot. Default: Maximum number of UMIs.}

\item{xlim}{Numeric vector with the upstream and downstream limits of the region to be plotted. Default: Maximum range in the trend object.}

\item{plot_raw}{Logical indicating wheter to plot the raw UMI-4C counts as points.}
}
\description{
Plot single trend
}
