% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot adaptative smoothen trend}
\usage{
plotTrend(
  umi4c,
  grouping = "condition",
  colors = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the \code{UMI4C} constructor.}

\item{grouping}{Name of the \code{colData(UMI4C)} column that will be used for grouping the profiles. If it is
vector of column names, will use both columns for grouping the samples.}

\item{colors}{Vector of colors to use for plotting for each group.}

\item{xlim}{Limits for the plot x axis (genomic coordinates).}

\item{ylim}{Limits of the trend y axis.}
}
\value{
Produces the adaptative trend plot, showing average UMIs at each position taking into
account the minimum number of molecules used to merge restriction fragments.
}
\description{
Plot adaptative smoothen trend
}
