% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMI4C.R
\name{plotUMI4C}
\alias{plotUMI4C}
\title{Plot UMI4C data}
\usage{
plotUMI4C(
  umi4c,
  dgram_function = "quotient",
  dgram_plot = TRUE,
  colors = NULL,
  xlim = NULL,
  ylim = NULL,
  TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene,
  longest = TRUE,
  rel_heights = c(0.25, 0.4, 0.12, 0.23),
  font_size = 14
)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the
\code{UMI4C} constructor.}

\item{dgram_function}{The name of the function used for calculating the
fold-change in the domainogram plot,
either "difference" or "quotient".}

\item{dgram_plot}{Logical indicating whether to plot the domainogram. If the
\code{umi4c} object only contains one sample will be automatically disabled.}

\item{colors}{Vector of colors to use for plotting for each group.}

\item{xlim}{Limits for the plot x axis (genomic coordinates).}

\item{ylim}{Limits of the trend y axis.}

\item{TxDb}{TxDb object to use for drawing the genomic annotation.}

\item{longest}{Logical indicating whether to plot only the longest
transcripts for each gene in the gene annotation part.}

\item{rel_heights}{Numeric vector of length 3 indicating the relative heights
of each part of the UMI4C plot.}

\item{font_size}{Base font size to use for the UMI4C plot.}
}
\value{
Produces a summary plot with all the information contained in the
UMI4C opject.
}
\description{
Produce a full UMI4C data plot containing the genes in the region, the
adaptative smoothen trend and the domainogram.
}
\examples{
umi <- makeUMI4Cexample()
plotUMI4C(umi,
          dgram_plot=FALSE)

}
