% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepR.R
\name{prepR}
\alias{prepR}
\title{umi4Cats prep}
\usage{
prepR(raw_dir, wk_dir, bait_seq, bait_pad, res_e)
}
\arguments{
\item{raw_dir}{Path with the fastq (compressed or descompressed)}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{res_e}{Restriction enzyme sequence}
}
\description{
Prepare the fastq files for being analyse by the umi4Cats pipeline filtering reads with bait and
introducing the respective UMI in the file headers
}
\examples{
raw_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/raw'
wk_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/MLH1_ctrl_umi4cats_python'
threads = 1
bait_seq = 'ACCTAGAAGGATATGCGCTTGC'
bait_pad = 'GCGTTAGA'
res_e = "GATC"

prepR(raw_dir, wk_dir, bait_seq, bait_pad, res_e)

}
