% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{prepUMI4C}
\alias{prepUMI4C}
\title{Prepare UMI4C data}
\usage{
prepUMI4C(fastq_dir, wk_dir, bait_seq, bait_pad, res_enz)
}
\arguments{
\item{fastq_dir}{Path of the directory containing the FastQ files (compressed or uncompressed).}

\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{bait_seq}{Character containing the bait primer sequence.}

\item{bait_pad}{Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).}

\item{res_enz}{Character containing the restriction enzyme sequence.}
}
\description{
Prepare the FastQ files for the analysis by selecting reads with bait and
adding the respective UMI identifier for each read in its header.
}
\examples{
\dontrun{
prepUMI4C(fastq_dir="raw_fastq",
      wk_dir="SOCS1",
      bait_seq="CCCAAATCGCCCAGACCAG",
      bait_pad="GCGCG",
      res_enz="GATC")
}

}
\seealso{
\code{\link{contactsUMI4C}}
}
