% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{processUMI4C}
\alias{processUMI4C}
\title{Process a single UMI4C profile}
\usage{
processUMI4C(UMI4C, scales = 5:150, min_win_cov = NULL,
  factor = 0.025, sd = 2)
}
\arguments{
\item{UMI4C}{UMI4C objected created with the \code{\link{UMI4C}} function.}

\item{scales}{Scales to use for smoothing and plotting the domainogram. Scales indicate how many fragment
ends are merged together to obtain the normalized number of UMIs. Default: 5:150.}

\item{min_win_cov}{Minimum number of UMIs requiered in a region to perform the adaptaive smoothing. Default: NULL,
number is calculated by the function \code{\link{getMinWinMols}} using the factor specified below.}

\item{factor}{Factor representing the proportion of UMIs from the total UMIs in the analyzed region needed as a minimum
to perform the adaptative smoothing.}

\item{sd}{Standard devition value for the adaptative smoothing. Default: 2 (extracted from Schwartzman et al).}
}
\value{
This function returns a list object with the following fields:
\itemize{
 \item \strong{raw}: A data.frame including the raw UMIs in each fragment end considered for the analysis
 (see \code{\link{filterRaw}}).
 \item \strong{bait}: A GRanges object containing the coordinates of the bait used in the UMI-4C assay.
 \item \strong{dgram}: A matrix containing the domainogram (see \code{\link{createDomainogram}}).
 \item \strong{trend}: A data.frame containing the adaptative smoothen trend for the assay
 (see \code{\link{smoothTrendAdaptative}}).
 }
}
\description{
This function processes a single UMI4C profile and generates all data necessary for plotting and performing differential analysis.
}
