% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{processUMI4C}
\alias{processUMI4C}
\title{Process a single UMI4C profile}
\usage{
processUMI4C(input, bait_coordinates, scales = 5:150,
  bait_exclusion = 3000, bait_upstream = 5e+05,
  bait_downstream = 5e+05, min_win_cov = NULL, factor = 0.025,
  sd = 2)
}
\arguments{
\item{input}{A two-column data.frame where the first column represents the fragend and the second the number of UMIs.}

\item{scales}{Scales to use for smoothing and plotting the domainogram. Default: 1:150.}

\item{bait_exclusion}{Region around the bait (in bp) to be excluded from the analysis. Default: 3kb.}

\item{bait_upstream}{Number of bp upstream of the bait to use for the analysis. Default: 500kb.}

\item{bait_downstream}{Number of bp downstream of the bait to use for the analysis. Default: 500kb.}

\item{min_win_cov}{Minimum number of UMIs requiered in a region to perform the adaptaive smoothing. Default: NULL,
number is calculated by the function getMinWinMols() using the factor specified below.}

\item{factor}{Factor representing the proportion of UMIs from the total UMIs in the analyzed region needed as a minimum
to perform the adaptative smoothing.}

\item{sd}{Standard devition value for the adaptative smoothing. Default: 2 (extracted from Schwartzman et al).}

\item{bait_coordintes}{GRanges object with the coordinates of the bait used for the experiment.}
}
\description{
This function processes a single UMI4C profile and generates all data necessary for plotting.
}
