% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processUMI4C.R
\name{smoothTrendAdaptative}
\alias{smoothTrendAdaptative}
\title{Adaptative smoothing of the UMI-4C trend}
\usage{
smoothTrendAdaptative(dgram, min_win_cov = NULL, bait_coordinates,
  factor = 0.025, sd = 2)
}
\arguments{
\item{dgram}{Output of the function \code{\link{createDomainogram}}.}

\item{min_win_cov}{Minimum number of UMIs requiered in a region to perform the adaptaive smoothing. Default: NULL,
number is calculated by the function \code{\link{getMinWinMols}} using the factor specified below.}

\item{bait_coordinates}{GRanges object with the coordinates of the bait used for the experiment.}

\item{factor}{Factor representing the proportion of UMIs from the total UMIs in the analyzed region needed as a minimum
to perform the adaptative smoothing.}

\item{sd}{Standard devition value for the adaptative smoothing. Default: 2 (extracted from Schwartzman et al).}
}
\value{
A data.frame containing the necessary information to draw the different elements of the UMI-4C trend:
\itemize{
\item \strong{coord}: Coordinates of the fragment end.
\item \strong{trend}: Smoothed UMIs value according to min_win_cov.
\item \strong{scale}: Scale used (# of fragends to merge) to achieve min_win_cov.
\item \strong{group}: Fragment ends grouped according to postion using bait as reference (upstream or downstream).
\item \strong{devP}: Upper bound of standard deviation.
\item \strong{devM}: Lower bound of standard deviation.
}
}
\description{
Performs adaptative smoothing to plot a UMI-4C contact trend. This means that joins fragment ends together until the
minimum number of UMIs in the joined fragment is >= min_win_cov.
}
