% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{split}
\alias{split}
\title{Split fastq files at a given restriction site}
\usage{
split(fastq_file, res_enz, cut_pos, split_dir, min_flen = 20, numb_reads)
}
\arguments{
\item{fastq_file}{Fastq file path.}

\item{res_enz}{Sequence for the restriction enzyme to cut.}

\item{cut_pos}{Position where RE cuts.}

\item{split_dir}{Directory where to save split files.}

\item{numb_reads}{Size of the sample (number of FASTQ reads) load on each loop.}
}
\description{
Split fastq files at a given restriction site
}
