% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{splitR}
\alias{splitR}
\title{Split UMI4C reads}
\usage{
splitR(wk_dir, res_enz, cut_pos, path_venv)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{cut_pos}{Numeric indicating the nucleotide position where restriction enzyme cuts (zero-based) (for example, for DpnII is 0).}

\item{path_venv}{Path for the UMI4Cats virtual environment generated using the \code{\link{installVenv}}.}
}
\description{
Split the prepared reads using the restrition enzyme information.
}
\examples{
\dontrun{
splitR(wk_dir="SOCS1",
       res_enz="GATC",
       cut_pos=0,
       path_venv="~/venvs/UMI4Cats_venv/")
}
}
