% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitR.R
\name{splitR}
\alias{splitR}
\title{umi4Cats splitting}
\usage{
splitR(py_umi4cats, wk_dir, res_e, cut_pos)
}
\arguments{
\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{res_e}{Restriction enzyme sequence}

\item{cut_pos}{Nucleotide where restriction enzyme cuts (zero-base)}
}
\description{
Split the prepared reads using the pre-defined restrition enzyme information
}
