% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{splitScript}
\alias{splitScript}
\title{Umi4c splitting}
\usage{
splitScript(wk_dir, cores, re, trimmomatic = "trimmomatic")
}
\arguments{
\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{trimmomatic}{Trimmomatic path. By default trimmomatic}
}
\description{
This function splits the filtered reads by the restriction enzyme defined.
}
