% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{splitUMI4C}
\alias{splitUMI4C}
\title{Split UMI4C reads}
\usage{
splitUMI4C(wk_dir, prep_dir = "", res_enz = "GATC", cut_pos = 0)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{cut_pos}{Numeric indicating the nucleotide position where restriction enzyme cuts (zero-based) (for example, for DpnII is 0).}
}
\description{
Split the prepared reads using the restrition enzyme information.
}
\examples{
\dontrun{
splitUMI4C(wk_dir="SOCS1",
       cut_seq_5p="",
       cut_seq_3p="GATC")
}
}
