% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogsUmi4Cats.R
\name{statsUMI4C}
\alias{statsUMI4C}
\title{Statistics UMI4C}
\usage{
statsUMI4C(fastq_dir, wk_dir)
}
\arguments{
\item{fastq_dir}{Path of the directory containing the FastQ files (compressed or uncompressed).}

\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}
}
\description{
Creates a stats file and generates a summary plot describing interesting statistics
for processed UMI-4C samples.
}
\examples{
\dontrun{
statsUMI4C(fastq_dir="raw_fastq",
           wk_dir="SOCS1")
}
}
