% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{UMI4C}
\alias{UMI4C}
\alias{.UMI4C}
\alias{UMI4C-class}
\title{UMI4C-class}
\arguments{
\item{colData}{Data.frame containing the information for constructing the UMI4C experiment object. Needs
to contain the following columns:
\itemize{
    \item sampleID Unique identifier for the sample.
    \item condition Condition for performing differential analysis. Can be control and treatment, two different cell types, etc.
    \item replicate Number for identifying replicates.
    \item file File as outputed by \code{umi4CatsContacts} function.
}}

\item{metadata}{List containing the following elements:
\enumerate{
\item bait: GRanges object representing the position of the bait used for the analysis.
}}

\item{assayData}{Matrix containing the ID for the restriction fragment as row.names and each column representing
the number of raw UMIs for a specific sample.}
}
\description{
UMI4C-class
}
\note{
The \code{UMI4C} object extends the \code{DESeqDataSet} class.
}
\keyword{package}
