% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialNbinomWaldTestUMI4C.R
\name{UMI4C2dds}
\alias{UMI4C2dds}
\title{UMI4Cats object to DDS object.}
\usage{
UMI4C2dds(umi4c, design = ~condition)
}
\arguments{
\item{umi4c}{UMI4C object as generated by \code{makeUMI4C} or the
\code{UMI4C} constructor.}

\item{design}{A \code{formula} or \code{matrix}. The formula expresses how
the counts for each fragment end depend on the variables in \code{colData}.
See  \code{\link[DESeq2]{DESeqDataSet}}.}
}
\value{
DDS object.
}
\description{
Transforms an UMI4C object to a DDS object
function.
}
