% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi4CatsContacts.R
\name{umi4CatsContacts}
\alias{umi4CatsContacts}
\title{Umi4c analysis}
\usage{
#'umi4CatsContacts(raw_dir, wk_dir, threads, bait_seq, bait_pad, res_e, ref_gen, genomic_track)
}
\arguments{
\item{raw_dir}{Path with the fastq (compressed or descompressed)}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{threads}{Number of threads to use in the analysis}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{res_e}{Restriction enzyme sequence}

\item{ref_gen}{Reference genome in fasta format}

\item{genomic_track}{Genomic track path generated by the generateGenomicTrack function of UMI4Cats package}

\item{cut_pos}{Nucleotide where restriction enzyme cuts (zero-base)}
}
\description{
This function parses, clean, splits, aligns and analyses fastq files counting chromatin contacts between the viewpoint and
the surrounding loci. The output is a tsv file used for the generation of 4C profile.
}
\examples{
raw_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/raw'
wk_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/MLH1_ctrl_umi4cats_python'
threads = 1
bait_seq = 'ACCTAGAAGGATATGCGCTTGC'
bait_pad = 'GCGTTAGA'
res_e = "GATC"
cut_pos = "0" # should be a string for properly working
ref_gen = '/imppc/labs/lplab/share/marc/refgen/hg19/hg19.fa'
genomic_track = '/imppc/labs/lplab/share/marc/epimutations/processed/genomicTracks/genomic_tracks_hg19/dpnII_genomicTrack'
umi4CatsContacts(raw_dir,
                 wk_dir,
                 threads,
                 bait_seq,
                 bait_pad,
                 res_e,
                 ref_gen,
                 genomic_track)

}
