% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi4CatsContacts.R
\name{umi4CatsContacts}
\alias{umi4CatsContacts}
\title{Umi4c analysis}
\usage{
umi4CatsContacts(input, cores, bait_seq, bait_pad, re, ref_gen, wk_dir,
  genomic_track, trimmomatic = "trimmomatic", bowtie2 = "bowtie2",
  fastqmultx = "fastq-multx")
}
\arguments{
\item{input}{Path with the fastq (compressed or descompressed)}

\item{cores}{Number of cores to use in the analysis}

\item{bait_seq}{Bait sequence.}

\item{re}{Restriction enzyme sequence.}

\item{ref_gen}{Reference genome in fasta format.}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis.}

\item{genomic_track}{Genomic track path}

\item{trimmomatic}{Trimmomatic path}

\item{bowtie2}{Bowtie2 path}

\item{fastqmultx}{Fastq-multx pàth}

\item{bait_path}{Bait pad.}
}
\description{
This function parses, splits, aligns and analyses fastq files counting chromatin contacts between the viewpoint and 
the surrounding loci.
}
