% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umi4catsVenvInst.R
\name{umi4catsVenvInst}
\alias{umi4catsVenvInst}
\title{Install python virtual environment}
\usage{
umi4catsVenvInst(pathVenv)
}
\arguments{
\item{pathVenv}{Path where to install the virtualenv}
}
\description{
Automatically install python virtual environment from requirements.txt.
}
\examples{
\dontrun{
}
}
