% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactsUMI4C.R
\name{umiCounterR}
\alias{umiCounterR}
\title{UMI counting}
\usage{
umiCounterR(wk_dir, bait_seq, bait_pad, res_enz, digested_genome, ref_gen,
  path_venv)
}
\arguments{
\item{wk_dir}{Working directory where to save the outputs generated by the UMI-4c analysis.}

\item{bait_seq}{Character containing the bait primer sequence.}

\item{bait_pad}{Character containing the pad sequence (sequence between the bait primer and the restriction enzyme sequence).}

\item{res_enz}{Character containing the restriction enzyme sequence.}

\item{digested_genome}{Path for the digested genome file generated using the \code{\link{digestGenome}} function.}

\item{ref_gen}{Path for the reference genome to use for the alignment (fasta format).}
}
\description{
Algorithm for counting  and collapsing the number of UMIs supporting a specific ligation.
}
\details{
For collapsing different molecules into the same UMI, takes into account the ligation position and
the number of UMI sequence mismatches.
}
\examples{
\dontrun{
umiCounterR(wk_dir="SOCS1",
            bait_seq="CCCAAATCGCCCAGACCAG",
            bait_pad="GCGCG",
            res_enz="GATC",
            digested_genome=hg19_dpnii,
            ref_gen="~/data/reference_genomes/hg19/hg19.fa",
            path_venv="~/venvs/UMI4Cats_venv/")
}
}
