% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiCounterR.R
\name{umiCounterR}
\alias{umiCounterR}
\title{umi4Cats umiCounter}
\usage{
umiCounterR(py_umi4cats, wk_dir, bowtie2, ref_gen, samtools, genomic_track,
  bait_seq, bait_pad, res_e)
}
\description{
@description Count number of UMIs for every ligation taking ligation position and UMI mismatches into acount. It generates a tsv file 
 for creating the 4c profile. 
 
 @usage umiCounterR(py_umi4cats, raw_dir, wk_dir, bait_seq, bait_pad, res_e, fastqmultx)
 
 @inheritParams umi4CatsContacts
}
