% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umiCounterR.R
\name{umiCounterR}
\alias{umiCounterR}
\title{umi4Cats umiCounter}
\usage{
umiCounterR(pathVenv, raw_dir, wk_dir, bait_seq, bait_pad, res_e, fastqmultx)
}
\arguments{
\item{pathVenv}{UMI4Cats virtual enviroment path}

\item{wk_dir}{Working directory where to carry out the UMI-4c analysis}

\item{ref_gen}{Reference genome in fasta format}

\item{genomic_track}{Genomic track path generated by the generateGenomicTrack function of UMI4Cats package}

\item{bait_seq}{Bait sequence}

\item{bait_pad}{Bait pad}

\item{res_e}{Restriction enzyme sequence}
}
\description{
Count number of UMIs for every ligation taking ligation position and UMI mismatches into acount. It generates a tsv file
for creating the 4c profile.
}
\examples{
\dontrun{
pathVenv = '/imppc/labs/lplab/share/marc/venv/umi4catsVenv'
wk_dir = '/imppc/labs/lplab/share/marc/epimutations/processed/prove/MLH1_ctrl_umi4cats_python'
bait_seq = 'ACCTAGAAGGATATGCGCTTGC'
bait_pad = 'GCGTTAGA'
res_e = "GATC"
ref_gen = '/imppc/labs/lplab/share/marc/refgen/hg19/hg19.fa'
genomic_track = '/imppc/labs/lplab/share/marc/epimutations/processed/genomicTracks/genomic_tracks_hg19/dpnII_genomicTrack'

umiCounterR(wk_dir, ref_gen, genomicTrack, bait_seq, bait_pad,res_e)

}

}
