% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialNbinomWaldTestUMI4C.R
\name{ZscoreUMI4C}
\alias{ZscoreUMI4C}
\title{Z-score calculation using residuals of trend and fit UMI4C counts}
\usage{
ZscoreUMI4C(dds)
}
\arguments{
\item{dds}{DDS object as generated by \code{smoothMonotoneUMI4C} with the 
smooth monotone fit counts}
}
\value{
DDS object with zscore,pvalue and padjusted assays
}
\description{
Calculates the z-score and then they are converted into one-sided P-values and 
adjusted for multiple testing using the method of Benjamini and Hochberg
}
\details{
This function calculates the z-score for each fragment over all samples from the residuals of the 
symmetric monotone fit and the median absolute deviation (mad). Z-scores are then converted 
into one-sided P-values using the standard Normal cumulative distribution function, 
and these are adjusted for multiple testing using the method of Benjamini and Hochberg
}
